- NEW: Removed the old DataImportingDialog code. - CHG: Some very minor changes to the map importing code to reduce memory usage. - NEW: Updated and changed some of the i18n text used during importing.

+++ src/flapjack/io/ChromosomeMapImporter.java
@@ -121,13 +121,13 @@ public class ChromosomeMapImporter
 class MarkerIndex
 {
 	// The index of the chromosome map within the data set
-	public int mapIndex;
+	public short mapIndex;
 	// The index of the marker itself within the map
 	public int mkrIndex;
 
 	public MarkerIndex(int mapIndex, int mkrIndex)
 	{
-		this.mapIndex = mapIndex;
+		this.mapIndex = (short) mapIndex;
 		this.mkrIndex = mkrIndex;
 	}
 }
